UPDATE:
XMLGenerator: I finished the header skeleton and roughed out some pseudo-code in the cpp... well, i started to and then quit because I got sick of thinking about it.  But I think its a pretty good approach to things.

WordIndex:  I added a public getter to WordNode so that the XMLGenerator can have access to each WordNode's OccurrenceSet.  Otherwise we would have to be its friend or something and thats crazy talk... but we need to iterate through the OccurrenceSet when we want to print out each occurrence.  

RENAME: 
	So I am an idiot and cant spell Occurrence... I spelled it "occurence" and had to do a project-wide find and replace to fix the misspelling (turned out to be quick and easy with XCode.... thanks apple.)


TODO: 
	-Still need to write the HTML parser and figure out how the HTMLTokenizer works... 
	-Need to finish my WebCrawler->Crawl() method.... maybe tomorrow